all old test work. before new tests

+++ src/main/java/ru/ivan/sqlcmd/controller/command/DropAllTables.java
@@ -4,6 +4,9 @@ import ru.ivan.sqlcmd.model.DatabaseManager;
 import ru.ivan.sqlcmd.view.View;
 
 public class DropAllTables extends Command {
+    private static final String ANSI_RED = "\u001B[31m";
+    private static final String ANSI_RESET = "\u001B[0m";
+
     @Override
     public String description() {
         return "drop all tables";
@@ -30,7 +33,18 @@ public class DropAllTables extends Command {
     @Override
     public void process(String command) {
 
-        manager.dropAllTables();
-        view.write("Все таблицы были успешно удалены");
+        confirmAndDropAllTables();
+    }
+
+    private void confirmAndDropAllTables() {
+        try {
+            view.write(ANSI_RED + "Удаляем все таблицы? Y/N" + ANSI_RESET);
+            if (view.read().equalsIgnoreCase("y")) {
+                manager.dropAllTables();
+                view.write("Все таблицы были успешно удалены.");
+            }
+        } catch (Exception e) {
+            view.write(String.format("Ошибка удаления всех таблиц по причине: %s", e.getMessage()));
+        }
     }
 }
